Feature add: show clipboard copy to stdout too
ogcopy always swallowed hole the input and did not show any of it. Like this it will also show if the user would like to see with just -s or --show options. Also added an option to remove the last new line of the piped input to avoid having weird copying